로그인 관련 파일명을 auth로 변경
